Moved overwrite translations to a separate test file
